Extract ObjectOrErrorLogProcessor module
