+++ lib/line-builders.js
@@ -1,37 +1,6 @@
 'use strict'
 
-const {
-  prettifyErrorLog,
-  prettifyObject
-} = require('./utils')
-
-function appendObjectOrError (lineParts, { log, EOL, IDENT, errorLikeObjectKeys, errorProps, messageKey }) {
-  const errorProperties = errorProps.split(',')
-
-  if (log.type === 'Error' && log.stack) {
-    const prettifiedErrorLog = prettifyErrorLog({
-      log,
-      errorLikeKeys: errorLikeObjectKeys,
-      errorProperties,
-      ident: IDENT,
-      eol: EOL
-    })
-    lineParts.push(prettifiedErrorLog)
-  } else {
-    const skipKeys = typeof log[messageKey] === 'string' ? [messageKey] : undefined
-    const prettifiedObject = prettifyObject({
-      input: log,
-      skipKeys,
-      errorLikeKeys: errorLikeObjectKeys,
-      eol: EOL,
-      ident: IDENT
-    })
-    lineParts.push(prettifiedObject)
-  }
-}
-
 const builtInLineBuilders = {
-  appendObjectOrError
 }
 
 function buildLine (lineBuilders, context) {
