+++ lib/log-processors.js
@@ -1,4 +1,3 @@
-const { builtInLineBuilders } = require('./line-builders')
 const LogProcessor = require('./LogProcessor')
 const JsonLogProcessor = require('./processors/JsonLogProcessor')
 const PrimitivesLogProcessor = require('./processors/PrimitivesLogProcessor')
@@ -10,10 +9,7 @@ const MetadataLogProcessor = require('./processors/MetadataLogProcessor')
 const SemicolonLogProcessor = require('./processors/SemicolonLogProcessor')
 const TimeLogProcessor = require('./processors/TimeLogProcessor')
 const EOLLogProcessor = require('./processors/EOLLogProcessor')
-
-const {
-  appendObjectOrError
-} = builtInLineBuilders
+const { ObjectOrErrorLogProcessor } = require('./processors/ObjectOrErrorLogProcessor')
 
 function createLogProcessor (definition) {
   if (definition instanceof LogProcessor) {
@@ -27,12 +23,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class ObjectOrErrorLogProcessor extends LogProcessor {
-  build (lineParts, context) {
-    return appendObjectOrError(lineParts, context)
   }
 }
